add a third argument to get the package scope pjson
