+++ lib/internal/modules/esm/resolve.js
@@ -356,7 +356,6 @@ function invalidPackageTarget(
 
 const invalidSegmentRegEx = /(^|\\|\/)((\.|%2e)(\.|%2e)?|(n|%6e|%4e)(o|%6f|%4f)(d|%64|%44)(e|%65|%45)(_|%5f)(m|%6d|%4d)(o|%6f|%4f)(d|%64|%44)(u|%75|%55)(l|%6c|%4c)(e|%65|%45)(s|%73|%53))?(\\|\/|$)/i;
 const deprecatedInvalidSegmentRegEx = /(^|\\|\/)((\.|%2e)(\.|%2e)?|(n|%6e|%4e)(o|%6f|%4f)(d|%64|%44)(e|%65|%45)(_|%5f)(m|%6d|%4d)(o|%6f|%4f)(d|%64|%44)(u|%75|%55)(l|%6c|%4c)(e|%65|%45)(s|%73|%53))(\\|\/|$)/i;
-const invalidPackageNameRegEx = /^\.|%|\\/;
 const patternRegEx = /\*/g;
 
 /**
@@ -752,44 +751,6 @@ function packageImportsResolve(name, base, conditions) {
   throw importNotDefined(name, packageJSONUrl, base);
 }
 
-/**
- * Parse a package name from a specifier.
- * @param {string} specifier - The import specifier.
- * @param {string | URL | undefined} base - The parent URL.
- */
-function parsePackageName(specifier, base) {
-  let separatorIndex = StringPrototypeIndexOf(specifier, '/');
-  let validPackageName = true;
-  let isScoped = false;
-  if (specifier[0] === '@') {
-    isScoped = true;
-    if (separatorIndex === -1 || specifier.length === 0) {
-      validPackageName = false;
-    } else {
-      separatorIndex = StringPrototypeIndexOf(
-        specifier, '/', separatorIndex + 1);
-    }
-  }
-
-  const packageName = separatorIndex === -1 ?
-    specifier : StringPrototypeSlice(specifier, 0, separatorIndex);
-
-  // Package name cannot have leading . and cannot have percent-encoding or
-  // \\ separators.
-  if (RegExpPrototypeExec(invalidPackageNameRegEx, packageName) !== null) {
-    validPackageName = false;
-  }
-
-  if (!validPackageName) {
-    throw new ERR_INVALID_MODULE_SPECIFIER(
-      specifier, 'is not a valid package name', fileURLToPath(base));
-  }
-
-  const packageSubpath = '.' + (separatorIndex === -1 ? '' :
-    StringPrototypeSlice(specifier, separatorIndex));
-
-  return { packageName, packageSubpath, isScoped };
-}
 
 /**
  * Resolves a package specifier to a URL.
@@ -804,57 +765,24 @@ function packageResolve(specifier, base, conditions) {
     return new URL('node:' + specifier);
   }
 
-  const { packageName, packageSubpath, isScoped } =
-    parsePackageName(specifier, base);
+  const { packageJSONUrl, packageJSONPath, packageSubpath } = packageJsonReader.getPackageJSONURL(specifier, base);
 
-  // ResolveSelf
-  const packageConfig = packageJsonReader.getPackageScopeConfig(base);
-  if (packageConfig.exists) {
-    if (packageConfig.exports != null && packageConfig.name === packageName) {
-      const packageJSONUrl = pathToFileURL(packageConfig.pjsonPath);
-      return packageExportsResolve(
-        packageJSONUrl, packageSubpath, packageConfig, base, conditions);
-    }
+  const packageConfig = packageJsonReader.read(packageJSONPath, { __proto__: null, specifier, base, isESM: true });
+
+  // Package match.
+  if (packageConfig.exports != null) {
+    return packageExportsResolve(
+      packageJSONUrl, packageSubpath, packageConfig, base, conditions);
+  }
+  if (packageSubpath === '.') {
+    return legacyMainResolve(
+      packageJSONUrl,
+      packageConfig,
+      base,
+    );
   }
 
-  let packageJSONUrl =
-    new URL('./node_modules/' + packageName + '/package.json', base);
-  let packageJSONPath = fileURLToPath(packageJSONUrl);
-  let lastPath;
-  do {
-    const stat = internalFsBinding.internalModuleStat(
-      internalFsBinding,
-      StringPrototypeSlice(packageJSONPath, 0, packageJSONPath.length - 13),
-    );
-    // Check for !stat.isDirectory()
-    if (stat !== 1) {
-      lastPath = packageJSONPath;
-      packageJSONUrl = new URL((isScoped ?
-        '../../../../node_modules/' : '../../../node_modules/') +
-        packageName + '/package.json', packageJSONUrl);
-      packageJSONPath = fileURLToPath(packageJSONUrl);
-      continue;
-    }
-
-    // Package match.
-    const packageConfig = packageJsonReader.read(packageJSONPath, { __proto__: null, specifier, base, isESM: true });
-    if (packageConfig.exports != null) {
-      return packageExportsResolve(
-        packageJSONUrl, packageSubpath, packageConfig, base, conditions);
-    }
-    if (packageSubpath === '.') {
-      return legacyMainResolve(
-        packageJSONUrl,
-        packageConfig,
-        base,
-      );
-    }
-
-    return new URL(packageSubpath, packageJSONUrl);
-    // Cross-platform root check.
-  } while (packageJSONPath.length !== lastPath.length);
-
-  throw new ERR_MODULE_NOT_FOUND(packageName, fileURLToPath(base), null);
+  return new URL(packageSubpath, packageJSONUrl);
 }
 
 /**
