+++ lib/internal/modules/package_json_reader.js
@@ -4,20 +4,26 @@ const {
   ArrayIsArray,
   JSONParse,
   ObjectDefineProperty,
+  RegExpPrototypeExec,
+  StringPrototypeIndexOf,
+  StringPrototypeSlice,
 } = primordials;
 const {
   fileURLToPath,
-  isURL,
   pathToFileURL,
   URL,
 } = require('internal/url');
-const { kEmptyObject } = require('internal/util');
 const {
-  validateString,
-  validateURLString,
-} = require('internal/validators');
+  codes: {
+    ERR_INVALID_MODULE_SPECIFIER,
+    ERR_MODULE_NOT_FOUND,
+  },
+} = require('internal/errors');
+const { kEmptyObject } = require('internal/util');
 const modulesBinding = internalBinding('modules');
 const path = require('path');
+const { validateString } = require('internal/validators');
+const internalFsBinding = internalBinding('fs');
 
 /**
  * @typedef {import('typings/internalBinding/modules').DeserializedPackageConfig} DeserializedPackageConfig
@@ -129,7 +135,7 @@ function readPackage(requestPath) {
 /**
  * Get the nearest parent package.json file from a given path.
  * Return the package.json data and the path to the package.json file, or undefined.
- * @param {URL['pathname']} checkPath The path to start searching from.
+ * @param {string} checkPath The path to start searching from.
  * @returns {undefined | DeserializedPackageConfig}
  */
 function getNearestParentPackageJSON(checkPath) {
@@ -180,38 +186,118 @@ function getPackageType(url) {
   return getPackageScopeConfig(url).type;
 }
 
+const invalidPackageNameRegEx = /^\.|%|\\/;
+/**
+ * Parse a package name from a specifier.
+ * @param {string} specifier - The import specifier.
+ * @param {string | URL | undefined} base - The parent URL.
+ */
+function parsePackageName(specifier, base) {
+  let separatorIndex = StringPrototypeIndexOf(specifier, '/');
+  let validPackageName = true;
+  let isScoped = false;
+  if (specifier[0] === '@') {
+    isScoped = true;
+    if (separatorIndex === -1 || specifier.length === 0) {
+      validPackageName = false;
+    } else {
+      separatorIndex = StringPrototypeIndexOf(
+        specifier, '/', separatorIndex + 1);
+    }
+  }
+
+  const packageName = separatorIndex === -1 ?
+    specifier : StringPrototypeSlice(specifier, 0, separatorIndex);
+
+  // Package name cannot have leading . and cannot have percent-encoding or
+  // \\ separators.
+  if (RegExpPrototypeExec(invalidPackageNameRegEx, packageName) !== null) {
+    validPackageName = false;
+  }
+
+  if (!validPackageName) {
+    throw new ERR_INVALID_MODULE_SPECIFIER(
+      specifier, 'is not a valid package name', fileURLToPath(base));
+  }
+
+  const packageSubpath = '.' + (separatorIndex === -1 ? '' :
+    StringPrototypeSlice(specifier, separatorIndex));
+
+  return { packageName, packageSubpath, isScoped };
+}
+
+function getPackageJSONURL(specifier, base) {
+  const { packageName, packageSubpath, isScoped } =
+    parsePackageName(specifier, base);
+
+  // ResolveSelf
+  const packageConfig = getPackageScopeConfig(base);
+  if (packageConfig.exists) {
+    if (packageConfig.exports != null && packageConfig.name === packageName) {
+      const packageJSONPath = packageConfig.pjsonPath;
+      return { packageJSONUrl: pathToFileURL(packageJSONPath), packageJSONPath, packageSubpath };
+    }
+  }
+
+  let packageJSONUrl =
+    new URL('./node_modules/' + packageName + '/package.json', base);
+  let packageJSONPath = fileURLToPath(packageJSONUrl);
+  let lastPath;
+  do {
+    const stat = internalFsBinding.internalModuleStat(
+      internalFsBinding,
+      StringPrototypeSlice(packageJSONPath, 0, packageJSONPath.length - 13),
+    );
+    // Check for !stat.isDirectory()
+    if (stat !== 1) {
+      lastPath = packageJSONPath;
+      packageJSONUrl = new URL((isScoped ?
+        '../../../../node_modules/' : '../../../node_modules/') +
+        packageName + '/package.json', packageJSONUrl);
+      packageJSONPath = fileURLToPath(packageJSONUrl);
+      continue;
+    }
+
+    // Package match.
+    return { packageJSONUrl, packageJSONPath, packageSubpath };
+  } while (packageJSONPath.length !== lastPath.length);
+
+  throw new ERR_MODULE_NOT_FOUND(packageName, fileURLToPath(base), null);
+}
+
 const pjsonImportAttributes = { __proto__: null, type: 'json' };
 let cascadedLoader;
 /**
- * @param {URL['href'] | URL['pathname']} specifier The location for which to get the "root" package.json
- * @param {URL['href'] | URL['pathname']} parentLocation The location of the current module (ex file://tmp/foo.js).
+ * @param {URL['href'] | string | URL} specifier The location for which to get the "root" package.json
+ * @param {URL['href'] | string | URL} parentLocation The location of the current module (ex file://tmp/foo.js).
  */
-function findPackageJSON(specifier, parentLocation) {
-  if (!isURL(specifier)) { validateString(specifier, 'specifier'); }
-
-  if (
-    !isURL(parentLocation) &&
-    !path.isAbsolute(parentLocation)
-  ) {
-    validateURLString(parentLocation, 'parentLocation');
+function findPackageJSON(specifier, parentLocation, getPackageScope = true) {
+  try {
+    specifier = `${specifier}`;
+  } catch {
+    validateString(specifier, 'specifier');
   }
 
+  if (typeof parentLocation === 'symbol') {
+    validateString(parentLocation, 'parentLocation');
+  }
   const parentURL = URL.parse(parentLocation) || pathToFileURL(parentLocation);
 
-  let resolvedTarget;
-  if (specifier[0] === '.') {
-    resolvedTarget = new URL(specifier, parentURL);
-  } else {
-    cascadedLoader ??= require('internal/modules/esm/loader').getOrInitializeCascadedLoader();
+  if (getPackageScope) {
+    const { packageJSONPath } = getPackageJSONURL(specifier, parentURL);
+    return packageJSONPath;
+  }
 
-    try {
-      resolvedTarget = cascadedLoader.resolve(specifier, `${parentURL}`, pjsonImportAttributes).url;
-    } catch (err) {
-      if (err.code === 'ERR_UNSUPPORTED_DIR_IMPORT') {
-        resolvedTarget = err.url;
-      } else {
-        throw err;
-      }
+  let resolvedTarget;
+  cascadedLoader ??= require('internal/modules/esm/loader').getOrInitializeCascadedLoader();
+
+  try {
+    resolvedTarget = cascadedLoader.resolve(specifier, `${parentURL}`, pjsonImportAttributes).url;
+  } catch (err) {
+    if (err.code === 'ERR_UNSUPPORTED_DIR_IMPORT') {
+      resolvedTarget = err.url;
+    } else {
+      throw err;
     }
   }
 
@@ -226,5 +312,6 @@ module.exports = {
   getNearestParentPackageJSON,
   getPackageScopeConfig,
   getPackageType,
+  getPackageJSONURL,
   findPackageJSON,
 };
