+++ lib/internal/validators.js
@@ -499,16 +499,6 @@ function validateUnion(value, name, union) {
   }
 }
 
-/**
- * @param {*} value
- * @param {string} name
- * @returns {asserts value is a parsable URL string}
- */
-const validateURLString = hideStackFrames((value, name) => {
-  const { URL } = require('internal/url'); // Can't be at the top
-  if (!URL.canParse(value)) throw new ERR_INVALID_ARG_TYPE(name, 'a URL string', value);
-});
-
 /*
   The rules for the Link header field are described here:
   https://www.rfc-editor.org/rfc/rfc8288.html#section-3
@@ -649,7 +639,6 @@ module.exports = {
   validateUint32,
   validateUndefined,
   validateUnion,
-  validateURLString,
   validateAbortSignal,
   validateLinkHeaderValue,
   validateInternalField,
